20251012-Receipt_Visualisation_E2E_Expense_Analytics/requirements.txt
anthropic>=0.39.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
pillow>=10.0.0
watchdog>=3.0.0
python-dotenv>=1.0.0